Refactor multisignature, to use index based.

+++ blockchain/src/config.rs
@@ -27,7 +27,7 @@ use serde_derive::{Deserialize, Serialize};
 #[derive(Debug, Clone)]
 pub struct BlockchainConfig {
     /// Maximal number of slots for election.
-    pub max_slot_count: usize,
+    pub max_slot_count: i64,
     /// Minimal stake amount.
     pub min_stake_amount: i64,
     /// Time to lock stakes.
